requirements.txt
flask>=2.3,<4
tensorflow>=2.15
tensorflow-metal; sys_platform == "darwin" and platform_machine == "arm64"
numpy>=1.23
pillow>=10
gunicorn>=21
reportlab>=4
matplotlib>=3.7
scikit-learn>=1.3
